[Security] moves tokens management in AuthenticationBundle
