+++ src/main/authentication/Resources/config/services.yml
@@ -7,3 +7,4 @@ imports:
     - { resource: services/security.yml }
     - { resource: services/serializer.yml }
     - { resource: services/updater.yml }
+    - { resource: services/voter.yml }
