+++ src/main/authentication/Resources/config/services/controller.yml
@@ -2,6 +2,17 @@ services:
     _defaults:
         public: true
 
+    Claroline\AuthenticationBundle\Controller\ApiTokenController:
+        parent: Claroline\AppBundle\Controller\AbstractCrudController
+        public: true
+        arguments:
+            - '@security.authorization_checker'
+            - '@security.token_storage'
+
+    Claroline\AuthenticationBundle\Controller\IpUserController:
+        parent: Claroline\AppBundle\Controller\AbstractCrudController
+        public: true
+
     Claroline\AuthenticationBundle\Controller\OauthController:
         parent: Claroline\AppBundle\Controller\AbstractCrudController
         public: true
@@ -9,7 +20,3 @@ services:
             - '@router'
             - '@Claroline\CoreBundle\Library\Configuration\PlatformConfigurationHandler'
             - '@Claroline\AuthenticationBundle\Manager\OauthManager'
-
-    Claroline\AuthenticationBundle\Controller\IpUserController:
-        parent: Claroline\AppBundle\Controller\AbstractCrudController
-        public: true
