+++ src/main/authentication/Resources/config/services/voter.yml
@@ -1,0 +1,15 @@
+services:
+    _defaults:
+        autowire: false
+        autoconfigure: false
+        public: true
+
+    _instanceof:
+        Claroline\CoreBundle\Security\Voter\AbstractVoter:
+            tags: [security.voter]
+            calls:
+                - method: setContainer
+                  arguments:
+                      - '@service_container'
+
+    Claroline\AuthenticationBundle\Security\Voter\ApiTokenVoter: ~
