+++ src/main/core/Resources/config/services/controller.yml
@@ -7,13 +7,6 @@ services:
           - '@Claroline\AppBundle\Persistence\ObjectManager'
           - '@Claroline\CoreBundle\Library\RoutingHelper'
 
-    Claroline\CoreBundle\Controller\APINew\Cryptography\ApiTokenController:
-        parent: Claroline\AppBundle\Controller\AbstractCrudController
-        public: true
-        arguments:
-          - '@security.authorization_checker'
-          - '@security.token_storage'
-
     Claroline\CoreBundle\Controller\APINew\Log\LogConnectController:
         arguments:
           - '@security.authorization_checker'
